global: move events and managers to their own namespace
