+++ src/main/java/event/timer/StudyTimerCanceledEvent.java
@@ -1,4 +1,4 @@
-package model.timer.event;
+package event.timer;
 
 import model.Course;
