+++ src/main/java/viewcontroller/CourseMainPage.java
@@ -18,8 +18,8 @@ import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.*;
+import manager.CourseManager;
 import model.Course;
-import model.CourseManager;
 import model.Moment;
 import model.ToDo;
