+++ src/main/java/viewcontroller/CoursePanelItem.java
@@ -11,7 +11,7 @@ public class CoursePanelItem {
   @FXML private Label courseName;
   @FXML private Label courseCode;
 
-  // TODO model.CourseManager & CourseFactory?
+  // TODO manager.CourseManager & CourseFactory?
   private Course course;
   private MainPage parentController;
